Moved and renamed a few things
