+++ firefly/src/components/MyAccount/ProfileCard.js
@@ -1,13 +1,13 @@
 import React, {useContext} from "react";
-import "../styles/MyAccountPage.scss";
+import "../../styles/MyAccountPage.scss";
 import { Link } from "react-router-dom";
 
-import Icon from "../assets/icons";
+import Icon from "../../assets/icons";
 import { FaPen } from "react-icons/fa";
 
 //context
-import { childContext } from "../context/ChildProfiles/ChildProfileStore";
-import { UPDATE_SELECTED } from "../context/ChildProfiles/ChildProfileStore";
+import { childContext } from "../../context/ChildProfiles/ChildProfileStore";
+import { UPDATE_SELECTED } from "../../context/ChildProfiles/ChildProfileStore";
 
 const ProfileCard = (props) => {
