+++ firefly/src/components/SignInAndSignUp/SignUpPage.js
@@ -4,12 +4,12 @@ import firebase from "firebase";
 import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth";
 import * as Yup from "yup";
 import ApolloClient from "apollo-boost";
-import WearingNerdGlasses from "./../images/WearingNerdGlasses.png";
-import firebaseApp from "./FirebaseLogin";
+import WearingNerdGlasses from "./../../images/WearingNerdGlasses.png";
+import firebaseApp from "../FirebaseLogin";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faEye } from "@fortawesome/free-solid-svg-icons";
 
-import "./../styles/SignUpPage.scss";
+import "./../../styles/SignUpPage.scss";
 
 const { gql } = require("apollo-boost");
