+++ firefly/src/styles/LoggedOutStartPage.scss
