+++ firefly/src/utils/buttons/PrimaryButton.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../styles/Button.scss";
+import "../../styles/Button.scss";
 
 export class PrimaryButton extends React.Component {
   render() {
