+++ firefly/src/utils/buttons/SecondaryButton.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../styles/Button.scss";
+import "../../styles/Button.scss";
 
 export class SecondaryButton extends React.Component {
   render() {
